processor/requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
langchain>=0.1.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pgvector>=0.2.4
